Refactor Vuex store and tests for admin dashboard after banging my head against a wall with Joe for a couple of hours
