+++ app/javascript/admin/dashboard/store/__mocks__/index.js
@@ -1,6 +1,8 @@
 import VuexMockStore from 'utilities/vuex-mock-store'
 
-import { state, getters, mutations } from '../index'
+import state from '../state'
+import getters from '../getters'
+import mutations from '../mutations'
 
 export default new VuexMockStore({
   state,
