+++ app/javascript/admin/dashboard/store/state.js
@@ -1,0 +1,4 @@
+export default {
+  chartEndpoints: {},
+  cachedStates: {},
+}
